[skip ci]: WIP: add progress bar component
